pyproject.toml
[tool.poetry]
name = "multiagent-chat"
version = "0.1.0"
description = "Multi-agent medical chat application with A2A protocol support and FHIR integration"
authors = ["OMRS AI Team <team@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"  # A2A SDK requires Python 3.10+
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.0.0"
python-multipart = "^0.0.6"
psutil = "^5.9.0"
requests = "^2.32.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
a2a-sdk = {extras = ["http-server"], version = "^0.3.2"}  # A2A SDK with HTTP server support
PyYAML = "^6.0"
PyHive = {version = "^0.7.0", optional = true}
thrift = {version = "^0.20.0", optional = true}
thrift_sasl = {version = "^0.4.3", optional = true}
pure-sasl = {version = "^0.6.2", optional = true}
duckdb = {version = "^0.8.0", optional = true}

[tool.poetry.extras]
spark = ["PyHive", "thrift", "thrift_sasl", "pure-sasl"]
duckdb = ["duckdb"]

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88